requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow-macos==2.16
tensorflow-metal==1.1